Валидация полей категорий, продуктов по правилам

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -38,17 +39,32 @@ class CategoryController extends Controller
 	 * @param  \Illuminate\Http\Request  $request
 	 * @return \Illuminate\Http\Response
 	 */
-	public function store(Request $request)
+	public function store(CategoryRequest $request)
 	{
-		// получим путь к файлу, который был загружен (в форме это input c type = "file" name = "image") с указанием имени 
-		// папки в которой картинка будет сохранена (+ч.13: Storage)
-		$path = $request->file('image')->store('categories');
+		// в обработку запроса с формы добавим валидацию (ч.14: Валидация, FormRequest):
+		/* $request->validate([
+
+			'code' => 'required',
+			'name' => 'required',
+			'description' => 'required',
+		]); */
 
 		// сохраним запрос на сохранение категории в переменной
 		$params = $request->all();
 
-		// изменем данные в ячейке: image на полученные
-		$params['image'] = $path;
+		unset($params['image']);
+
+		// для необязательного поля добавим проверку при получении запроса:
+		if ($request->has('image')) {
+
+			// получим путь к файлу, который был загружен (в форме это input c type = "file" name = "image") с указанием имени 
+			// папки в которой картинка будет сохранена (+ч.13: Storage)
+			$path = $request->file('image')->store('categories');
+
+			// если есть картинка, то добавим её в параметры (в соответствующую ячейку)
+			$params['image'] = $path;
+		}
+
 
 		Category::create($params);
 
@@ -84,16 +100,22 @@ class CategoryController extends Controller
 	 * @param  \App\Models\Category  $category
 	 * @return \Illuminate\Http\Response
 	 */
-	public function update(Request $request, Category $category)
+	public function update(CategoryRequest $request, Category $category)
 	{
-		// чтобы при редактировании категории загрузить новую картинку, старую нужно удалить (+ч.13: Storage)
-		Storage::delete($category->image);
-
-		$path = $request->file('image')->store('categories');
-
 		$params = $request->all();
 
-		$params['image'] = $path;
+		unset($params['image']);
+
+		// если в запросе с формы редактирования категории картинка есть, то мы должны её добавить, если нет, то мы недолжны старую удалять
+		if ($request->has('image')) {
+
+			// чтобы при редактировании категории загрузить новую картинку, старую нужно удалить (+ч.13: Storage)
+			Storage::delete($category->image);
+
+			$path = $request->file('image')->store('categories');
+
+			$params['image'] = $path;
+		}
 
 		$category->update($params);
